pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdapi"
version = "0.1.0"
description = "A Geometry Dash API Wrapper"
authors = [
    { name="notanerd1", email="ngoca0309@outlook.com" }
]
license = { text = "MIT" }
dependencies = [
    "httpx",
    "aiofiles",
    "lxml",
    "attrs",
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Clients",
    "Development Status :: Alpha"
]
[project.urls]
homepage = "https://github.com/notanerd314/gdapi"
source = "https://github.com/notanerd314/gdapi"
tracker = "https://github.com/notanerd314/gdapi/issues"

[project.scripts]
requires-python = ">=3.8,<4.0"

[tool.setuptools]
packages = ["gd"]

[tool.setuptools.package-data]
"gd" = ["*"]